新增进度调节方式

+++ sakatimeline/src/main/java/io/github/rangaofei/sakatimeline/TimeLineView.java
@@ -180,7 +180,7 @@ public class TimeLineView extends RecyclerView {
         initData();
     }
 
-    public void updateDividerNum(int dividerNum, boolean showAnim) {
+    public void updateDividerNum(float dividerNum, boolean showAnim) {
         ExceptionUtil.checkIfNull(this.timeLineConfig, "TimeLineConfig");
         ExceptionUtil.checkIfNull(this.timeLineConfig.getStepViewConfig(), "StepViewConfig");
 
@@ -193,7 +193,7 @@ public class TimeLineView extends RecyclerView {
 
     }
 
-    public void updateDividerNum(int dividerNum) {
+    public void updateDividerNum(float dividerNum) {
         updateDividerNum(dividerNum, true);
     }
 }
